refactor: distribute type definitions to their respective modules
- Moved Config and CommandConfig types to ConfigManager.ts
- Moved ExecutionRequest and ExecutionResult types to CommandExecutor.ts
- Moved LogPaths type to LogWriter.ts
- Updated all imports to use types from their new locations
- Deleted types/index.ts as it's no longer needed
- Each module now owns its own type definitions, improving cohesion
