+++ src/config/ConfigManager.test.ts
@@ -1,7 +1,7 @@
 import { promises as fs } from "node:fs";
 import * as path from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
-import type { Config } from "../types";
+import type { Config } from "./ConfigManager";
 import {
   getAvailableKeys,
   getCommand,
